schedule/requirements.txt
ariadne==0.26.2
Flask==3.1.2
grpcio==1.75.0
grpcio-tools==1.75.0
requests==2.32.5
dotenv
pymongo==4.15.3